+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- `moops.embed.Passthrough`, a utility for reusing the result of one embedded
+  notebook as input to another.
+
 ## [0.3.0] - 2026-05-13
 
 ### Added